crm114-20080326-BlameSentansoken-Ger-2249
2008-05-01 Ger Hobbelt

    * The error handling routines now 'know' when we're inside the
      debugger, so failures in watched expressions (which can easily
      happen ;-) ) won't have side-effects such as execution of trap
      handlers in the script itself (ouch!).

    * Improved the debugger and updated QUICKREF.txt accordingly.

    * Debugger 'v' command: line range handling split off to a separate
      routine; now handles a more flexible format too (every number can
      be 'defaulted').

    * Did we now finally stamp out all the boundary bugs in expandvar()?
      -- This following a bug report by Jason Lewis, which showed a
      boundary bug in vanilla CRM114, but upon renewed inspection I
      found the GerH copy still could also suffer from boundary issues
      (though is fewer, rather obscure conditions, yet a boundary bug is
      one too many anyhow).

    * expandvar() code cleanup - at least a bit.

      KNOWN BUG: ":@: :@: 1 + 1 ::" is not handled properly by
                 expandvar: nested match expressions in an 'eval'
                 statement screw up BADLY.

      TODO: for performance reasons, the quad loop through the buffer
            for each expansion should be reduced to a single round only.
            This can be nicely combined with the fix for KNOWN BUG
            above.

    * debugger: 'c' restored to its proper glory: 'c' without any arg
      should just GO. We've got 'n' for single stepping, don't we?

    * makefiles/configure: upgraded to the latest autoconf/automake/etc.

      KNOWN BUGS: some of the AX_... macros use undocumented
                  AC_RUN_LOG() which does not exist anymore in the
                  latest release.

                  This must be fixed (diff commandline flag checks and a
                  few others suffer from this)

    * makefiles: 'make check' target now does something VERY useful: it
      runs all the megatest tests (except Neural Net) as separate tests.
      See tests/Makefile.am for the individual make targets: use these
      to test only a subset of the available tests -- very handy when
      debugging a certain bit of code or classifier.

      This is another step towards a complete 'standardized' make/conf
      setup:

        ./bootstrap      <-- for maintainers only!

        ./configure

        make

        make check

        make install

        make clean

      and for maintainers the extra:

        make distcheck

        make dist

        make distclean

    * KNOWN BUG: 'make distcheck' does not really like my 'alternative'
                 .tar.gz distro archive naming scheme, so it barfs after
                 running most of the tests.

                 This needs to be fixed.

    * Win32: KNOWN BUG: syscall() is b0rked: generally it'll work, but
             there are certain spurious conditions related to the OS and
             environment which can cause the current code to hang
             forever.

             This should be fixed by moving to the new fully async
             hnadling scheme -- not ready for release that yet. :-(

    * compiler checks (table in crm_compiler.c) have been corrected for
      some of the more modern (experimental) classifiers.

    * makefiles: if you want a verbose 'make check', i.e. when you want
      to debug the scripts, do so using the environment variable:

         CRM114_MAKE_SCRIPTS_DEBUG=1; export CRM114_MAKE_SCRIPTS_DEBUG;

    * makefiles/tests: a new testscript.sh[.in] test driver has been
      added. This driver is used to run the individual tests; several
      specialized postprocessing filter scripts (using SED and AWK)
      having been added too, so results of a test can be validated in a
      flexible way. For example, see this classifier test which takes
      certain floating point issues into account upon auto-
      comparison/validation:

        Support_Vector_Machine_Unigram_test1: $(TEST_PREREQUISITES) test_timing_seconds_announcement
            $(E) "*******************************************"
            $(E) "* Support Vector Machine (SVM) unigram classifier "
            $(E) "*******************************************"
            [...]
            $(SILENT)\
            CRM114_CHECK_FILTER_ARGS="-v prob=0.1 -v pR=0.2"; export CRM114_CHECK_FILTER_ARGS; \
            ./testscript.sh '-{ isolate (:s:); {classify < svm unigram unique > ( i_test.css | q_test.css | i_vs_q_test.css ) (:s:) /[[:graph:]]+/ /0 0 100 1e-3 1 0.5 1 1/ [:_dw:]   ; output / type I \n:*:s:\n/} alius { output / type Q \n:*:s:\n/ } }' \
                $(srcdir)/mt_ng_Support_Vector_Machine_SVM_unigram_1.input \
                $(refdir)/Support_Vector_Machine_Unigram_test1.step4.refoutput \
                $(builddir)/crm114_classify.filter.sh
            [...]

      Note the CRM114_CHECK_FILTER_ARGS environment variable in there
      which sets custom AWK variables to configure the custom AWK/SH
      postprocess filter script 'crm114_classify.filter.sh' to tolerate
      a certain spread in 'prob' probability and 'pR' values. See also
      the 'crm114_classify.filter.awk' AWK script which does the
      postprocessing.

      The concept
      -----------

      testscript.sh is the main driver: it starts CRM114 with the
      specified script (either file or 'literal' '-{...}' script code;
      feed CRM114 any specified input file/literal text over stdin and
      store the stdout and stderr output in temporary files, which are
      merged with the CRM114 exit code into a template file, which is
      fed, together with its equivalent reference file from the
      tests/ref/ directory, to the postprocessing section: this is
      either a custom script wich compares both files, or standard UNIX
      diff. Any differences found are reported to the console and will
      cause cause the test to 'FAIL'. A match with the reference file
      will produce and 'OK' instead.

      The only custom postprocessing/diff script at this date which
      accepts extra parameters is crm114_classify.filter.sh, which gets
      its extra args through aforementioned environment variable.

      The other custom scripts (tests/ref/*.filter.sh) serve particular
      purposes and don't require parameterization.

    * TODO: tests/Makefile.am cleanup: now some of those lines are
            _extremely_ large; these may be broken into readable bits.

+++ src/crm_expr_fscm.c
@@ -651,8 +651,8 @@ static int deflate(SCM_STATE_STRUCT *s, char *t, int len, int *starts, int
 		fatalerror_ex(SRC_LOC(), "FSCM tries to deflate %d data words to a global buffer which is too small (%d words). "
 			"Please run crm114 with the '-w %d' (or a larger number) commandline switch.",
 				len,
-				data_window_size / sizeof(int),
-				len * sizeof(int));
+				(int)(data_window_size / sizeof(int)),
+				(int)(len * sizeof(int)));
 		return -1;
 	}
